feat: song menu sheet
