+++ src/components/commonui/bottomsheet/SheetIconButton.tsx
@@ -1,4 +1,4 @@
-import { View } from 'react-native';
+import { StyleSheet, View } from 'react-native';
 import { Text, IconButton } from 'react-native-paper';
 
 import { useNoxSetting } from '@stores/useApp';
@@ -21,24 +21,30 @@ export default ({
   const playerStyle = useNoxSetting(state => state.playerStyle);
 
   return (
-    <View
-      style={{
-        flex: 1,
-        height: 110,
-        alignItems: 'center',
-      }}
-    >
+    <View style={styles.view}>
       <IconButton
         disabled={disabled}
         icon={icon}
         mode={'contained-tonal'}
         size={30}
-        style={{ flex: 1, width: '90%' }}
+        style={styles.button}
         onPress={onPress}
         containerColor={playerStyle.colors.surface}
       />
-      {text && <Text>{text}</Text>}
+      {text && <Text style={disabled && styles.disabledText}>{text}</Text>}
       {children}
     </View>
   );
 };
+
+const styles = StyleSheet.create({
+  view: {
+    flex: 1,
+    height: 110,
+    alignItems: 'center',
+  },
+  button: { flex: 1, width: '90%' },
+  disabledText: {
+    color: 'gray',
+  },
+});
