+++ src/components/songmenu/SongMenuSheet.tsx
@@ -111,7 +111,7 @@ export default () => {
         }}
       >
         <Image
-          source={song?.cover}
+          source={{ uri: song?.cover, width: 50, height: 50 }}
           style={{ width: 50, height: 50, borderRadius: 5 }}
         />
         <View style={{ paddingLeft: 5, marginTop: -10 }}>
